changed the nav bar based on whether the user authorized or not
